Transactions module progress

+++ src/main/java/org/example/pruebatecnica/aplicacion/product/ProductServiceImpl.java
@@ -3,6 +3,8 @@ package org.example.pruebatecnica.aplicacion.product;
 import org.example.pruebatecnica.dominio.client.Client;
 import org.example.pruebatecnica.dominio.product.Product;
 import org.example.pruebatecnica.dominio.product.ProductService;
+import org.example.pruebatecnica.dominio.product.ProductStatus;
+import org.example.pruebatecnica.dominio.product.ProductType;
 import org.example.pruebatecnica.dominio.transaction.Transaction;
 import org.example.pruebatecnica.infraestructura.producto.ProductRepositoryImpl;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -18,13 +21,37 @@ public class ProductServiceImpl implements ProductService {
     private ProductRepositoryImpl productRepositoryImpl;
 
     @Override
-    public Product createProduct(Client clientId, String type, String status, Integer number, Integer balance, Boolean gmf) {
+    public Product createProduct(Client clientId, ProductType type, Integer balance, Boolean gmf) {
+        if (clientId == null) {
+            throw new IllegalArgumentException("El producto se debe asociar a un cliente");
+        }
+
+        ProductStatus status = ProductStatus.ACTIVA;
+
+        String generatedNumber = generateAccountNumber(type);
         List<Transaction> transactions = new ArrayList<>();
-        return productRepositoryImpl.save(new Product(clientId, type, status, number, balance, gmf, transactions));
+        return productRepositoryImpl.save(new Product(clientId, type, status, generatedNumber, balance, gmf, transactions));
     }
 
+    private String generateAccountNumber(ProductType type) {
+        String prefix = (type == ProductType.CUENTA_DE_AHORROS) ? "53" : "33";
+        return prefix + generateRandomDigits(8);
+    }
+
+    private String generateRandomDigits(int numberOfDigits) {
+        // Lógica para generar una cadena de dígitos numéricos aleatorios
+        Random random = new Random();
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < numberOfDigits; i++) {
+            sb.append(random.nextInt(10));
+        }
+        return sb.toString();
+    }
+
+
+
     @Override
-    public Product updateProduct(Long productId, String type, String status, Integer number, Integer balance, Boolean gmf) {
+    public Product updateProduct(Long productId, ProductType type, ProductStatus status, String number, Integer balance, Boolean gmf) {
         Product existingProduct = productRepositoryImpl.findById(productId);
         if (existingProduct != null) {
 
@@ -42,12 +69,19 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public String deleteProduct(Long productId) {
         Product existingProduct = productRepositoryImpl.findById(productId);
+
         if (existingProduct != null) {
-            productRepositoryImpl.delete(existingProduct);
-            return "Cliente eliminado";
+            if (existingProduct.getBalance() == 0) {
+                productRepositoryImpl.delete(existingProduct);
+                return "Cuenta eliminada";
+            } else {
+                return "No se puede eliminar la cuenta porque tiene un saldo diferente de $0";
+            }
         }
-        return "Cliente no existente";
+
+        return "Cuenta no existente";
     }
+
 
     @Override
     public List<Product> getAllProductsByClientId(Long clientId) {
